minor: edit format and parameter data type changing

+++ src/Particle/LoadBalance/Par_LB_CollectParticle2OneLevel.cpp
@@ -281,8 +281,8 @@ void Par_LB_CollectParticle2OneLevel( const int FaLv, const long AttBitIdx, cons
 
 
 // 1-4. set the array offsets of the send buffer of each patch (mainly for the OpenMP parallelization)
-   int *OffsetEachPatch_Patch   = new int [NPatchAll];
-   int *OffsetEachPatch_ParData = new int [NPatchAll];   // actually useless in the JustCountNPar mode
+   int  *OffsetEachPatch_Patch   = new int  [NPatchAll];
+   long *OffsetEachPatch_ParData = new long [NPatchAll];   // actually useless in the JustCountNPar mode
 
    for (int lv=FaLv+1, q=0; lv<=MAX_LEVEL; lv++, q++)
    for (int PID=0; PID<amr->NPatchComma[lv][1]; PID++)
@@ -294,8 +294,8 @@ void Par_LB_CollectParticle2OneLevel( const int FaLv, const long AttBitIdx, cons
 //    skip patches without particles (must skip since OffsetEachRank_* are not set for patches without particles)
       if ( NParThisPatch == 0 )  continue;
 
-      OffsetEachPatch_Patch  [AccIdx] = OffsetEachRank_Patch  [TRank];
-      OffsetEachPatch_ParData[AccIdx] = OffsetEachRank_ParData[TRank];
+      OffsetEachPatch_Patch  [AccIdx] =       OffsetEachRank_Patch  [TRank];
+      OffsetEachPatch_ParData[AccIdx] = (long)OffsetEachRank_ParData[TRank];
 
       OffsetEachRank_Patch  [TRank] += 1;
       OffsetEachRank_ParData[TRank] += NParThisPatch*NAtt;
@@ -324,9 +324,9 @@ void Par_LB_CollectParticle2OneLevel( const int FaLv, const long AttBitIdx, cons
             Aux_Error( ERROR_INFO, "OffsetEachPatch_Patch[%d] (%d) >= max (%d) !!\n",
                        AccIdx, OffsetEachPatch_Patch[AccIdx], NSendPatchTotal );
 
-         if ( OffsetEachPatch_ParData[AccIdx] + NAtt > NSendParTotal*NAtt )
-            Aux_Error( ERROR_INFO, "OffsetEachPatch_ParData[%d] + NAtt (%d) > max (%d) !!\n",
-                       AccIdx, OffsetEachPatch_ParData[AccIdx] + NAtt, NSendParTotal*NAtt );
+         if ( OffsetEachPatch_ParData[AccIdx] + (long)NAtt > NSendParTotal*(long)NAtt )
+            Aux_Error( ERROR_INFO, "OffsetEachPatch_ParData[%d] + NAtt (%ld) > max (%ld) !!\n",
+                       AccIdx, OffsetEachPatch_ParData[AccIdx] + (long)NAtt, NSendParTotal*(long)NAtt );
 #        endif
 
          SendBuf_NParEachPatch [ OffsetEachPatch_Patch[AccIdx] ] = NParThisPatch;
